Update server interface/port configuration to be done via flags.
Update finch server such that http and https server interface configuration flags allow for specifying interface an addition to port, as well as being able to turn either server off.

Unfortunately this is a breaking change to flags and the config object, though impact should be relatively low since finch server isn't explicitly coupled to the version of finch (or even twitter server for that matter as long as the interface is stable).